backend/requirements.txt
lightrag-hku>=1.4.9
numpy>=1.24.0
networkx>=3.0
openai>=1.0.0
google-generativeai>=0.3.0
yt-dlp>=2023.12.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0

